feat/ Add rough suggestion on compile and testing on server
Added skeleton on how to compile and test the cloned branch. Idea is
currently to execute a shell script using a process inside java that
compile, tests, writes to logs and echoes a message depending on
outcome. That message can then be compared to inside java where later
actions are performed. This outline is only a suggestion and will be
altered later!