feat/ Added compilation + log script and integrated with webhook (#25)
* feat/ Add primitive attempt at cloning repo inside Java

Added method containing outline for possible way to clone branch of
interest after receiving a pull notification from the webhook. Method
should be called from wrapper or similair probably.
Issues still exists that needs to be ironed out.

* feat/ Add rough suggestion on compile and testing on server

Added skeleton on how to compile and test the cloned branch. Idea is
currently to execute a shell script using a process inside java that
compile, tests, writes to logs and echoes a message depending on
outcome. That message can then be compared to inside java where later
actions are performed. This outline is only a suggestion and will be
altered later!

* fix/ Add line to wait for child process to return

Forgot to call process.waitFor() which should be included to block
thread until script is finished running.

* feat/ Improve script, remove redundant dependency

This commit:
    * Removes dependency on jgit by cloning inside script instead
    * Improves logging format of script from json outline to plain log
    * Script logs additional build information
    * Improves interaction with merged display methods by agreeing on
      common logs directory
    * Slight modification to merged display methods to make it
      clickable(might be moved to seperate branch if group decides on it)

Co-authored-by: Lussebullen <Lussebullen@users.noreply.github.com>

* feat/ Simplify decision making with exit codes

Since we do not have to report what went wrong script returns exit code
1 on failure. Further actions handled in java, to be implemented later.

* fix/ Integrated webhook with log system

Co-authored-by: Omar-AV <Omar-AV@users.noreply.github.com>
Integrated the compileTestLog.sh script to be triggered from the
github-webhook endpoint and pass exitcodes on to the commit-status
function.

---------

Co-authored-by: Omar Askar Vergara <omaraskar@outlook.com>
Co-authored-by: Lussebullen <Lussebullen@users.noreply.github.com>